refactor : ProgressRecord 에 필드를 JoinCars에서 List<Car>로 변경해서 joinCars와의 강결합 방지, 그 이유는 1. ProgressRecord 에 JoinCars를 인자로 불러와서 다시 내부에서 deepCopy하는 것인데 joinCars 를 꺼내쓰기(파고들기)보다 joinCars에서 제공해주는 방향이 보다 내부 구조를 보여주지 않아서 결합도가 낮아짐 2. JoinCars 가 필드가 아니게 됨으로서 JoinCars 자체를 파고들어서 특정 행동을 취하지 않아도 되게 됨
