+++ src/main/java/racinggame/ResultView.java
@@ -11,9 +11,9 @@ public class ResultView {
     public static void printAllResult(GameResponse results) {
         System.out.println("실행 결과");
         for (ProgressRecord record : results.gameResult().progressRecords()) { // 각 라운드
-            for (int i = 0; i < record.joinCars().cars().size() ; i++) {
-                System.out.print(record.joinCars().cars().get(i).getName() + " : ");
-                System.out.println(makeTrack(record.joinCars().cars().get(i).findLocation()));
+            for (int i = 0; i < record.cars().size() ; i++) {
+                System.out.print(record.cars().get(i).getName() + " : ");
+                System.out.println(makeTrack(record.cars().get(i).findLocation()));
             }
             System.out.println();
         }
