+++ src/test/java/racinggame/model/JoinCarsTest.java
@@ -18,32 +18,5 @@ class JoinCarsTest {
         
         assertThat(joinCars).isEqualTo(expectedJoinCars);
     }
-    
-    
-    @Test
-    void findJointWinnersTest() {
-        List<Car> joinCars = new ArrayList<>();
-        joinCars.add(new Car("pobi",5));
-        joinCars.add(new Car("crong",5));
-        joinCars.add(new Car("mo",3));
-        joinCars.add(new Car("jae",2));
-        JoinCars jc = new JoinCars(joinCars);
-        
-        WinnerCars expectedWc = new WinnerCars(List.of(new Car("pobi",5), new Car("crong",5)));
-        assertThat(jc.findWinners()).isEqualTo(expectedWc);
-    }
-    
-    @Test
-    void findSoleWinnersTest() {
-        List<Car> joinCars = new ArrayList<>();
-        joinCars.add(new Car("pobi",4));
-        joinCars.add(new Car("crong",5));
-        joinCars.add(new Car("mo",3));
-        joinCars.add(new Car("jae",2));
-        JoinCars jc = new JoinCars(joinCars);
-        
-        WinnerCars expectedJc = new WinnerCars(List.of(new Car("crong",5)));
-        assertThat(jc.findWinners()).isEqualTo(expectedJc);
-    }
-    
+
 }
