+++ src/test/java/racinggame/model/ProgressRecordTest.java
@@ -17,7 +17,7 @@ class ProgressRecordTest {
         joinCars.add(new Car("crong",5));
         joinCars.add(new Car("mo",3));
         joinCars.add(new Car("jae",2));
-        List<ProgressRecord> progressRecords = List.of(new ProgressRecord(new JoinCars(joinCars)));
+        List<ProgressRecord> progressRecords = List.of((new JoinCars(joinCars).toProcessRecord()));
         
         WinnerCars winner = new WinnerCars(List.of(new Car("pobi",5), new Car("crong",5)));
         
@@ -31,11 +31,10 @@ class ProgressRecordTest {
         joinCars.add(new Car("crong",5));
         joinCars.add(new Car("mo",3));
         joinCars.add(new Car("jae",2));
-        List<ProgressRecord> progressRecords = List.of(new ProgressRecord(new JoinCars(joinCars)));
+        List<ProgressRecord> progressRecords = List.of((new JoinCars(joinCars).toProcessRecord()));
         
         WinnerCars winner = new WinnerCars(List.of(new Car("crong",5)));
         
         assertThat(progressRecords.getLast().findGameWinners()).isEqualTo(winner);
     }
-
 }
